+++ src/Pipedrive.net/Models/Response/SimpleDeal.cs
@@ -19,7 +19,7 @@ namespace Pipedrive
         public DealStatus Status { get; set; }
 
         [JsonProperty("value")]
-        public long Value { get; set; }
+        public decimal Value { get; set; }
 
         [JsonProperty("currency")]
         public string Currency { get; set; }
@@ -38,5 +38,5 @@ namespace Pipedrive
 
         [JsonProperty("formatted_value")]
         public string FormattedValue { get; set; }
-	}
+    }
 }
